+++ Integrations/JiraV2/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [Unreleased]
 - API token parameter is now encrypted.
-- Improved authentication methods support (OAuth1, OAuth2, Basic)
+- Improved authentication methods support (OAuth1, OAuth2, Basic) (only from 5.0) 
 
 ## [19.8.2] - 2019-08-22
 - Add support for remote application links.